LRC done

+++ lrc_data.py
@@ -2,12 +2,15 @@ import os
 import argparse
 import re
 import time
+from os.path import exists
 
 max_k = 12
 max_parity = 4
 chunksize = 128
+mode = "j"
 throughput_filename = "throughput.log"
-output_filename = "data.log"
+javars_output_filename = "data/javars_lrc.csv"
+isa_l_output_filename = "data/isa-l_lrc.csv"
 
 def run_benchmark(k, l, r, p, mode):
     os.system(f"../run_benchmark.sh -k {k} -l {l} -r {r} -p {p} -c {chunksize} -m {mode} -f {throughput_filename} -e l")
@@ -33,8 +36,6 @@ def convertible(k, l, r):
 
 def generate_data(mode):
 
-    data = []
-
     if (mode == "i"):
         os.chdir("isa-l")
     elif (mode == "j"):
@@ -42,7 +43,6 @@ def generate_data(mode):
     else:
         print("ERROR: Incorrect mode\n")
         exit()
-        
     for k in range(1, max_k):
         for l in range(1, int(max_k / 2) + 1):
             for r in range(1, max_parity + 1):
@@ -51,39 +51,57 @@ def generate_data(mode):
                     # print("Error: LRC incovertable")
                     # exit()
                 for p in range(1, max_parity + 1):
+                    config_exists = False
+                    file_exists = exists("../" + output_file)
+                    if file_exists:
+                        with open("../" + output_file, "r") as f:
+                            lines = f.readlines()
+                            for line in lines:
+                                content = line.split(",")
+                                if (content[0] == str(k) and content[1] == str(l) and
+                                    content[2] == str(r) and content[3] == str(p)):
+                                    config_exists = True
+                                    break
+                    if config_exists:
+                        print(f"Data for configuration ({k}, {l}, {r}, {p}) already exists\n")
+                        continue
                     start_time = time.time()
                     print(f"Generating Data for: ({k}, {l}, {r}, {p})\n")
                     run_benchmark(k, l, r, p, mode)
                     throughput = point_throughput(mode)
                     os.remove(throughput_filename)
-                    data.append(((k, l, r, p), throughput))
+                    with open("../" + output_file, "a+") as f:
+                        f.seek(0)
+                        firstline = f.readline().rstrip()
+                        if firstline != "k,l,r,p,throughput":
+                            f.write("k,l,r,p,throughput\n")
+                        f.seek(0, 2)
+                        f.write(f"{k},{l},{r},{p},{throughput}\n")
                     print(f"Configuration: ({k}, {l}, {r}, {p})\tThroughput: {throughput}\n")
                     print("--- %s seconds elapsed in calculation ---\n" % (time.time() - start_time))
     os.chdir("..")
-    return data
-
-def write_data(data):
-    with open(output_filename, "w") as f:
-        for datum in data:
-            config, throughput = datum
-            k, l, r, p = config
-            print(f"Configuration: ({k}, {l}, {r}, {p})\tThroughput: {throughput}\n")
-            f.write(f"({k}, {l}, {r}, {p})\t{throughput}\n")
 
 def parse_args():
     global chunksize
     global output_file
+    global mode
     parser = argparse.ArgumentParser()
     parser.add_argument("-c", help="Chunksize in KB", default=chunksize, type=int)
-    parser.add_argument("-o", help="Output file name.", default=output_filename, type=str)
+    parser.add_argument("-m", help="Mode", default=mode, type=str)
     args = parser.parse_args()
     chunksize = args.c
-    output_file = "data/" + args.o
+    mode = args.m
+    if mode == "i":
+        output_file = isa_l_output_filename
+    elif mode == "j":
+        output_file = javars_output_filename
+    else:
+        print("ERROR: Incorrect mode\n")
+        exit()
 
 def main():
     parse_args()
-    data = generate_data("j")
-    write_data(data)
+    generate_data(mode)
 
 if __name__ == "__main__":
     main()
